refactor(model): model 디렉토리 구조 통일 (dto 디렉토리를 model 디렉토리로 이동)
