+++ src/main/java/com/balgoorm/balgoorm_backend/user/model/dto/request/UserSignupRequest.java
@@ -1,4 +1,4 @@
-package com.balgoorm.balgoorm_backend.user.dto.request;
+package com.balgoorm.balgoorm_backend.user.model.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
