+++ src/main/java/com/balgoorm/balgoorm_backend/user/model/dto/response/UserResponseDTO.java
@@ -1,4 +1,4 @@
-package com.balgoorm.balgoorm_backend.user.dto.response;
+package com.balgoorm.balgoorm_backend.user.model.dto.response;
 
 import com.balgoorm.balgoorm_backend.user.model.entity.User;
 import lombok.Getter;
@@ -6,19 +6,24 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 
 @Getter
-public class MyInfoResponseDTO {
+public class UserResponseDTO {
+    private Long id;
     private String userId;
+//    private String password;
     private String nickname;
     private String email;
     private LocalDateTime createDate;
+    private String role;
+//    private String boards;
 
-
-    public MyInfoResponseDTO(User user) {
-
+    public UserResponseDTO(User user) {
+        this.id = user.getId();
         this.userId = user.getUserId();
+//        this.password = user.getUserPassword();
         this.nickname = user.getNickname();
         this.email = user.getEmail();
         this.createDate = user.getCreateDate();
-
+        this.role = user.getRole().toString();
+//        this.boards = user.getBoards().toString();
     }
 }
